Added Movement and can now change screen size
Added some action listener-type things so the goomba can move around and also added a configuration so that we can change the size of the screen, looks quite a bit nicer. Tried to implement physics but I don't think 2DBox is installed on my system so ill look into that next.